Tighten class-load-time transform internals
The class-load-time transform's IOException catch was unreachable
because the synthetic entry it wraps never performs I/O; it
collapses into one helper. The synthetic entry is private to its
sole caller now, preventing accidental reuse. Two static factory
methods that only forwarded to a constructor with a trailing null
are inlined.
