--- platforms/core-runtime/classpath/src/main/java/org/gradle/internal/classpath/transforms/SyntheticClassEntry.java
@@ -1,60 +1,0 @@
-/*
- * Copyright 2026 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.internal.classpath.transforms;
-
-import org.gradle.api.file.RelativePath;
-import org.gradle.internal.classpath.ClasspathEntryVisitor;
-
-/**
- * Adapts a (className, bytes) pair to {@link ClasspathEntryVisitor.Entry} so that a
- * {@link ClassTransform} can be invoked outside of a JAR-walking context, such as at
- * class-load time.
- */
-public final class SyntheticClassEntry implements ClasspathEntryVisitor.Entry {
-
-    public static SyntheticClassEntry of(String className, byte[] content) {
-        return new SyntheticClassEntry(className, content);
-    }
-
-    private final String className;
-    private final byte[] content;
-
-    private SyntheticClassEntry(String className, byte[] content) {
-        this.className = className;
-        this.content = content;
-    }
-
-    @Override
-    public String getName() {
-        return className + ".class";
-    }
-
-    @Override
-    public RelativePath getPath() {
-        return RelativePath.parse(true, className + ".class");
-    }
-
-    @Override
-    public CompressionMethod getCompressionMethod() {
-        return CompressionMethod.UNDEFINED;
-    }
-
-    @Override
-    public byte[] getContent() {
-        return content;
-    }
-}
